Draft scripts for MYTH airdrop (#1)
The following scripts have been drafted:
* airdrop-generateBatches.js - Maps active users into the raw call data in the 643 directory, where the filename contains the call hash.  
* airdrop-asMulti.js - For all the files in the 643 directory, uses the "asMulti" to submit the batch.
* airdrop-approveAsMulti.js - For all the files in the 643 directory, uses the "approveAsMulti" to approve the batch.

Using the airdrop-generateBatches.js and the mythos-airdrop-v4-activeusers data from https://dune.com/substrate/mythos-airdrop-v4 we ran airdrop-generateBatches.js to populate the 643 directory with raw call data and filenames.  The filename contains the call hash (the hash of the call data), supporting verification trivially.

The files in this directory are used to drive airdrop-asMulti.js and airdrop-approveAsMulti.js which can be run by a member of a multisig (e.g. 2/3) and then approved by another member.